+++ src/wiki_loader/src/lib.rs
@@ -1,232 +1,27 @@
+
 // Standard Lib
 use std::fmt::{Debug, Display};
 use std::fs::File;
 use std::io::prelude::*;
-use std::io::{BufReader, BufWriter, SeekFrom};
+use std::io::{BufReader, SeekFrom};
 use std::path::{Path, PathBuf};
 
 // Third Party
 use bzip2::read::MultiBzDecoder;
-use fst::automaton::{Automaton, StartsWith, StartsWithState, Str, Subsequence};
-use fst::{IntoStreamer, Map, MapBuilder, Streamer};
 use quick_xml;
 use rayon::prelude::*;
 use serde::{Deserialize, Serialize};
 use serde_json;
 
+pub mod page;
+pub mod search;
+use crate::page::{DetailedPage, Page, SiteInfo};
+use crate::search::{Searcher, Searchable};
+
 fn create_directory_if_not_exists(path: &str) {
     let path = path.replace("~", std::env::var("HOME").unwrap().as_str());
     if !std::path::Path::new(&path).exists() {
         std::fs::create_dir_all(path).unwrap();
-    }
-}
-
-#[derive(Deserialize, Debug, Serialize)]
-struct NameSpace {
-    #[serde(rename = "@key")]
-    key: String,
-    #[serde(rename = "@case")]
-    case: String,
-    #[serde(rename = "$value")]
-    value: Option<String>,
-}
-
-#[derive(Deserialize, Debug, Serialize)]
-struct NameSpaces {
-    namespace: Vec<NameSpace>,
-}
-
-#[derive(Deserialize, Debug, Serialize)]
-#[serde(rename(serialize = "siteinfo"))]
-struct SiteInfo {
-    sitename: String,
-    dbname: String,
-    base: String,
-    generator: String,
-    case: String,
-    namespaces: Vec<NameSpaces>,
-}
-
-#[derive(Deserialize, Debug, Serialize, PartialEq, Eq, PartialOrd, Ord)]
-pub struct Redirect {
-    #[serde(rename = "@title")]
-    pub title: String,
-}
-
-#[derive(Deserialize, Debug, Serialize, PartialEq, Eq, PartialOrd, Ord)]
-pub struct Text {
-    #[serde(rename = "@bytes")]
-    pub bytes: u32,
-    #[serde(rename = "@xml:space")]
-    pub xml_space: Option<String>,
-    #[serde(rename = "$value")]
-    pub value: Option<String>,
-}
-
-#[derive(Deserialize, Debug, Serialize, PartialEq, Eq, PartialOrd, Ord)]
-pub struct RevisionDetailedPage {
-    pub id: u32,
-    pub parentid: Option<u32>,
-    pub timestamp: String,
-    pub format: Option<String>,
-    pub model: String,
-    pub text: Option<Text>,
-}
-
-#[derive(Deserialize, Debug, Serialize, PartialEq, Eq, PartialOrd, Ord)]
-pub struct RevisionPage {
-    pub id: u32,
-    pub parentid: Option<u32>,
-    pub timestamp: String,
-    pub format: Option<String>,
-    pub model: String,
-}
-
-#[derive(Deserialize, Debug, Serialize, PartialEq, Eq, PartialOrd, Ord)]
-pub struct Page {
-    pub title: String,
-    pub id: u32,
-    pub block_id: Option<usize>,
-    // Shouldn't be needed for indexing so saving memory
-    // pub ns: u32,
-    // pub redirect: Option<Redirect>,
-    // pub revision: Option<RevisionPage>,
-}
-
-#[derive(Deserialize, Debug, Serialize, PartialEq, Eq, PartialOrd, Ord)]
-pub struct DetailedPage {
-    pub title: String,
-    pub ns: u32,
-    pub id: u32,
-    pub block_id: Option<usize>,
-    pub redirect: Option<Redirect>,
-    pub revision: Option<RevisionDetailedPage>,
-}
-
-impl Display for DetailedPage {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let mut out = format!("Title:{},ns:{},id:{}", self.title, self.ns, self.id);
-        out = match &self.redirect {
-            Some(redirect) => format!("{},redirect:{}", out, redirect.title),
-            None => out,
-        };
-        return write!(f, "{}", out);
-    }
-}
-
-trait PageItem {}
-impl PageItem for Page {}
-impl PageItem for DetailedPage {}
-
-#[derive(Deserialize, Debug, Serialize)]
-#[serde(rename(serialize = "mediawiki"))]
-struct MediaWiki {
-    siteinfo: SiteInfo,
-    page: Vec<Page>,
-}
-
-pub trait Searchable {
-    fn new() -> Self;
-    fn get(&self, key: &str) -> Option<u64>;
-    fn len(&self) -> usize;
-    fn search(&self, query: &str) -> std::io::Result<Vec<(String, u64)>>;
-    fn open_searcher(&mut self, path: &str) -> std::io::Result<()>;
-    fn create_searcher(&mut self, pages: &Vec<Page>, output_path: &str) -> std::io::Result<()>;
-}
-
-#[derive(Debug)]
-pub struct Searcher {
-    map: Option<Map<Vec<u8>>>,
-}
-
-impl Searchable for Searcher {
-    fn new() -> Searcher {
-        let searcher = Searcher { map: None };
-        return searcher;
-    }
-
-    fn len(&self) -> usize {
-        if self.map.is_none() {
-            return 0;
-        }
-        let len = self.map.as_ref().unwrap().len();
-        return len;
-    }
-
-    fn get(&self, key: &str) -> Option<u64> {
-        if self.map.is_none() {
-            return None;
-        }
-        let val = self.map.as_ref().unwrap().get(key);
-        return val;
-    }
-
-    fn open_searcher(&mut self, path: &str) -> std::io::Result<()> {
-        self.map = Some(Map::new(std::fs::read(path).unwrap()).unwrap());
-        return Ok(());
-    }
-
-    fn create_searcher(&mut self, pages: &Vec<Page>, output_path: &str) -> std::io::Result<()> {
-        println!("Creating Searcher");
-        let mut key_val_tuple = pages
-            .iter()
-            .map(|page| {
-                let key = page.title.as_bytes();
-                let block_id = page.block_id.unwrap() as u64;
-                let page_id = page.id as u64;
-
-                // Store block_id and page_id (u32) in a u64
-                let value = (block_id << 32) | page_id;
-                (key, value)
-            })
-            .collect::<Vec<_>>();
-
-        key_val_tuple.sort_by(|a, b| a.0.cmp(b.0));
-        let mut wtr = BufWriter::new(File::create(output_path).unwrap());
-        let mut build = MapBuilder::new(&mut wtr).unwrap();
-
-        let mut set = std::collections::HashSet::new();
-
-        for (key, value) in key_val_tuple.iter() {
-            if set.contains(key) {
-                continue;
-            }
-            set.insert(key);
-            build.insert(key, *value).unwrap();
-        }
-
-        build.finish().unwrap();
-        let map = Map::new(std::fs::read(output_path).unwrap()).unwrap();
-
-        self.map = Some(map);
-        return Ok(());
-    }
-
-    fn search(&self, query: &str) -> std::io::Result<Vec<(String, u64)>> {
-        if self.map.is_none() {
-            return Err(std::io::Error::new(
-                std::io::ErrorKind::NotFound,
-                "Searcher data not initialised",
-            ));
-        }
-
-        let matcher = Str::new(query);
-
-        let mut results: Vec<(String, u64)> = Vec::new();
-
-        let mut stream = self
-            .map
-            .as_ref()
-            .unwrap()
-            .search_with_state(&matcher)
-            .into_stream();
-
-        while let Some((k, v, _s)) = stream.next() {
-            let datstr = String::from_utf8(k.to_vec()).unwrap();
-            results.push((datstr, v));
-        }
-
-        return Ok(results);
     }
 }
